Removed the use of doc.name
Earlier versions used the `doc.name` as the overall title of the document, but the discussion revealed that probably it is better to just trust the `doc_chunk.meta.headings` to have the title information sooner or later.  So I've removed all the special title stuff and am just relying on the headers now.

Signed-off-by: Bill Murdock <bmurdock@redhat.com>